feat: update styles and fonts for improved design
Add new font families iAWriter and Handjet to enhance typography. 
Adjust button styles by changing padding to margin and setting 
border width to zero for a cleaner look. Update logo class 
for consistency and correct the title in the HTML for accuracy. 
Refactor viewport dimensions to use a variable for better 
maintainability.